[opt](Nereids) refactor the PartitionTopN (#20102)
Do some small refactoring for the `PartitionTopN` and also address the left comment in #18784
